+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [ADD NEW VERSION HERE] - ADD DATE HERE
 ### Fixed
+ - Fix drivers compilation failur with kernel v6.15.
 ### Added
 ### Removed
  - Removed eBPF support, because I have no time to maintain it and it is not